Trim X509_INFO and move to crypto/pem
X509_INFO only exists to be a return value to PEM_X509_INFO_read. There
is no use in letting callers create these objects, since they cannot do
anything with it. Only X509_INFO_free is needed.

Also cut a ton of unused fields from X509_PKEY.

Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/64127
Commit-Queue: Bob Beck <bbe@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: David Benjamin <davidben@google.com>

(cherry picked from commit fcd464ce97d96fd0278ad3082a8429022ae2c4d7)
